fix: keep PromptHMR weights in fp32, use dtype only for autocast
The bundled SMPL-X package (smplx/lbs.py) has hardcoded float32
operations that can't coexist with bf16/fp16 model weights, so
casting the whole PromptHMR model to bf16 caused runtime dtype
mismatches inside lbs().

Change the semantics of LoadPromptHMR's dtype param: weights stay
in fp32, and dtype is used only for torch.autocast during forward.
This still gives the matmul/attention speedup of low-precision
compute without the model-weight casting failures.

Also revert the smpl_wrapper.py patch (no longer needed since
body_joints is back to fp32).

Separately, add timing logs to Sapiens PromptHMR Human Pose: prints
total PromptHMR and Sapiens inference time per run, plus ms/frame
and ms/crop, with CUDA synchronisation for accurate measurement.

+++ prompthmr_lib/prompt_hmr/smpl_family/smpl_wrapper.py
@@ -113,7 +113,7 @@ class SMPLX(smplx.SMPLXLayer):
         # Replace some SMPLX joints with SMPL joint for better alignment with datasets
         ### replace: elbows, knees and ankles
         smpl_joints = self.smpl_joints @ smplx_output.vertices
-        body_joints[:, [3,6,10,13,11,14]] = smpl_joints[:, [19,18,5,4,8,7]].to(body_joints.dtype)
+        body_joints[:, [3,6,10,13,11,14]] = smpl_joints[:, [19,18,5,4,8,7]].float()
         
         smplx_output.body_joints = body_joints
         smplx_output.face_joints = face_joints
